requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-multipart>=0.0.20
python-jose>=3.3.0
jinja2>=3.1.5
email-validator>=2.2.0
